Fix enemy spawn delay

+++ src/enemy/enemy.cpp
@@ -4,8 +4,8 @@
 #include <iostream>
 #include "../game/texturemanager.hpp"
 
-Enemy::Enemy(float max_hp, float speed, float x, float y, float size, int delay,
-             EnemyTypes type)
+Enemy::Enemy(float max_hp, float speed, float x, float y, float size,
+             float delay, EnemyTypes type)
     : max_hp_(max_hp),
       hp_(max_hp),
       speed_(speed),
@@ -65,7 +65,7 @@ void Enemy::Move(const std::vector<std::pair<int, int>>& path) {
 float Enemy::GetHp() const { return hp_; }
 float Enemy::GetMaxHp() const { return max_hp_; }
 float Enemy::GetSpeed() const { return speed_; }
-int Enemy::GetDelay() const { return delay_; }
+float Enemy::GetDelay() const { return delay_; }
 const std::pair<float, float> Enemy::GetPosition() const { return {x_, y_}; }
 const std::pair<int, int> Enemy::GetTile() const { return {int(x_), int(y_)}; }
 EnemyTypes Enemy::GetType() const { return type_; }
