+++ src/game/play_state.cpp
@@ -252,7 +252,7 @@ void PlayState::Tick() {
   auto cur_time = clock_.getElapsedTime().asSeconds();
   // Add enemies to the enemies vector with a certain delay
   if (spawn_queue_.size() > 0) {
-    int delay = spawn_queue_.front().GetDelay();
+    float delay = spawn_queue_.front().GetDelay();
     if (cur_time - last_spawn_ > delay && spawn_queue_.size() > 0) {
       enemies_.push_back(spawn_queue_.front());
       spawn_queue_.pop_front();
